Revert "[Refactor] Use MMCV MODEL_REGISTRY (#495)"
revert PR#495 to solve init_weight problem
