+++ mmdet3d/datasets/registry.py
@@ -1,0 +1,3 @@
+from mmcv.utils import Registry
+
+OBJECTSAMPLERS = Registry('Object sampler')
