+++ mmdet3d/ops/pointnet_modules/registry.py
@@ -1,0 +1,3 @@
+from mmcv.utils import Registry
+
+SA_MODULES = Registry('point_sa_module')
